+++ src/Command/RotateCommand.php
@@ -58,7 +58,7 @@ class RotateCommand extends Command
 
         try {
             //Gets deleted files
-            $files = (new BackupManager())->rotate((int)$args->getArgument('keep'));
+            $files = BackupManager::rotate((int)$args->getArgument('keep'));
 
             if (!$files) {
                 $io->verbose(__d('database_backup', 'No backup has been deleted'));