+++ src/Driver/Postgres.php
@@ -22,43 +22,4 @@ use DatabaseBackup\Driver\Driver;
  */
 class Postgres extends Driver
 {
-    /**
-     * Gets the value for the `--dbname` option for export and import
-     *  executables as string. It contains the connection string with username,
-     *  password and hostname.
-     *
-     * It returns something like:
-     * <code>
-     * postgresql://postgres@localhost/travis_ci_test
-     * </code>
-     * @return string
-     */
-    protected function getDbnameAsString(): string
-    {
-        return sprintf(
-            'postgresql://%s%s@%s/%s',
-            $this->getConfig('username'),
-            $this->getConfig('password') ? ':' . $this->getConfig('password') : '',
-            $this->getConfig('host'),
-            $this->getConfig('database')
-        );
-    }
-
-    /**
-     * Gets the executable command to export the database
-     * @return string
-     */
-    protected function _exportExecutable(): string
-    {
-        return sprintf('%s --format=c -b --dbname=%s', $this->getBinary('pg_dump'), $this->getDbnameAsString());
-    }
-
-    /**
-     * Gets the executable command to import the database
-     * @return string
-     */
-    protected function _importExecutable(): string
-    {
-        return sprintf('%s --format=c -c -e --dbname=%s', $this->getBinary('pg_restore'), $this->getDbnameAsString());
-    }
 }